+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## 20.11.0
+* !! Consent change !! To use remote config, you now need to give "remote-config" consent
+* !! Push breaking changes !! Google play vulnerability issue fixed due to broadcast receiver for android push notification
+* Added Surveys and NPS feedback widgets
+* Added replaceAllAppKeysInQueueWithCurrentAppKey method to replace all app keys in queue with the current app key
+* Added removeDifferentAppKeysFromQueue method to remove all different app keys from the queue
+* Added setStarRatingDialogTexts method to set text's for different fields of star rating dialog
+* Added recordAttributionID method to set the attribution ID for iOS.
+* Example app updated with single plugin for both IDFA and App tracking permission for iOS.
+* Device id NSNull check added for iOS to fix the length on null crash.
+* Added setLocationInit method to record Location before init, to prevent potential issues occurred when location is passed after init.
+* Added giveConsentInit method to give Consents before init, some features needed consent before init to work properly.
+* Fixed issues related to location tracking
+* Session stop and start safe checks added.
+* Updated underlying android SDK to 20.11.0
+* Updated underlying ios SDK to 20.11.0
+
 ## 20.04.1
 * Adding APM calls
 * Improved unhandled crash catching